Seems like done

+++ CourseWork/Bot.h
@@ -4,11 +4,11 @@
 class Bot {
 private:
 	Cell*** cellsRef;
-	short attacksWeights[5][2];
+	float attacksWeights[5][2];
 	Player player;
 
-	short checkLineAttack(Player subPlayer, sf::Vector2i currentCell, sf::Vector2i direction);
-	std::pair<short, sf::Vector2i> checkAttacks(Player subPlayer, sf::Vector2i cellToCheck);
+	float checkLineAttack(Player subPlayer, sf::Vector2i currentCell, sf::Vector2i direction);
+	std::pair<float, sf::Vector2i> checkAttacks(Player subPlayer, sf::Vector2i cellToCheck);
 
 public:
 	Bot(Cell*** cellsLink);
